Update evaluate_model_bias.py
Improved classification mechanism to use math for determining the likelihood of the category in the context of the sentence + question, for more accurate gpt2 classification. 

Added bert kcentroid quotes to the ensemble, this is the algorithm used in the grounding knn mechanism as well.

Further improved the mechanism to find the best quotes for half of the codebook from the summaries to determine the collective model interpretation for comparison of the ensemble with any of its parts.